+++ src/Service/Storage/StorageFactory.php
@@ -4,6 +4,8 @@ namespace Kuusamo\Vle\Service\Storage;
 
 class StorageFactory
 {
+    private static $provider;
+
     /**
      * Create a storage interface.
      *
@@ -11,6 +13,16 @@ class StorageFactory
      */
     public static function create(): StorageInterface
     {
-        return new LocalStorage;
+        return self::$provider ?? new LocalStorage;
+    }
+
+    /**
+     * Set a custom provider.
+     *
+     * @param StorageInterface $provider Storage provider.
+     */
+    public static function setProvider(StorageInterface $provider)
+    {
+        self::$provider = $provider;
     }
 }